+++ packages/Datadog.Unity/Runtime/DatadogSdk.cs
@@ -24,6 +24,7 @@ namespace Datadog.Unity
         private DatadogWorker _worker;
         private IInternalLogger _internalLogger = new PassThroughInternalLogger();
         private ResourceTrackingHelper _resourceTrackingHelper;
+        private GameObject _performanceTrackerObject;
 
         /// <summary>
         /// Gets the version of the SDK reported to Datadog. This strips the final (revision) part of the version.
@@ -73,7 +74,7 @@ namespace Datadog.Unity
         /// </summary>
         public static void Shutdown()
         {
-            Instance.ShutdownInstance();
+            Instance.ShutdownInstance(false);
         }
 
         /// <summary>
@@ -281,16 +282,55 @@ namespace Datadog.Unity
 
         private void EnableRum(DatadogConfigurationOptions options)
         {
-            if (options.RumApplicationId is null or "")
+            // Require an application ID
+            if (string.IsNullOrEmpty(options.RumApplicationId))
             {
                 _internalLogger.Log(DdLogLevel.Error, "Datadog RUM is enabled but an Application ID is not set.");
                 return;
             }
 
+            // Prepare our interface to the platform-specific RUM API, and register it to handle worker messages in a
+            // background thread
             var platformRum = _platform.InitRum(options);
             _worker.AddProcessor(DdRumProcessor.RumTargetName, new DdRumProcessor(platformRum));
-            Rum = new DdWorkerProxyRum(_worker);
 
+            // Create our main-thread IDdRum interface, which will enqueue messages for the worker-thread implementation
+            // to handle
+            var rumProxy = new DdWorkerProxyRum(_worker);
+            Rum = rumProxy;
+
+            // If we're configured to send performance metrics, instantiate a helper object that will observe frame
+            // timings from within the Unity scene
+            if (options.VitalsUpdateFrequency != VitalsUpdateFrequency.None)
+            {
+                // Create a persistent GameObject and add the DatadogPerformanceTracker behavior to it
+                _performanceTrackerObject = new GameObject("DatadogPerformanceTracker");
+                var performanceTracker = _performanceTrackerObject.AddComponent<DatadogPerformanceTracker>();
+                _performanceTrackerObject.hideFlags = HideFlags.HideAndDontSave;
+                UnityEngine.Object.DontDestroyOnLoad(_performanceTrackerObject);
+
+                // Determine how often we should read from the PerformanceTracker and pass the latest performance
+                // metrics into the platform SDK
+                float reportIntervalSeconds;
+                switch (options.VitalsUpdateFrequency)
+                {
+                    case VitalsUpdateFrequency.Frequent:
+                        reportIntervalSeconds = 0.1f;
+                        break;
+                    case VitalsUpdateFrequency.Average:
+                        reportIntervalSeconds = 0.5f;
+                        break;
+                    default:
+                        reportIntervalSeconds = 1.0f;
+                        break;
+                }
+
+                // Pass a reference to our performance tracker into the RUM proxy, and configure it to periodically
+                // sample the latest values (on the main thread) and initiate updateExternalRefreshRate calls
+                rumProxy.InitPerformanceTracker(performanceTracker, reportIntervalSeconds);
+            }
+
+            // Track scene changes as RUM Views, if configured to do so
             if (options.AutomaticSceneTracking)
             {
                 SceneManager.activeSceneChanged += SceneManagerOnActiveSceneChanged;
@@ -306,18 +346,33 @@ namespace Datadog.Unity
             });
         }
 
-        private void ShutdownInstance()
+        /// <summary>
+        /// Cleans up internal state owned by the SDK, either when quitting the application or when tearing down test
+        /// state.
+        /// </summary>
+        /// <param name="isQuitting">Whether we're being called in response to <c>Application.quitting</c>.</param>
+        private void ShutdownInstance(bool isQuitting)
         {
+            // Clean up any game objects added to the scene by the SDK, but avoid doing so if the application is
+            // shutting down: only delete objects in response to manual DatadogSdk.Shutdown() calls, i.e. in unit tests
+            if (!isQuitting && _performanceTrackerObject != null)
+            {
+                UnityEngine.Object.Destroy(_performanceTrackerObject);
+                _performanceTrackerObject = null;
+            }
+
+            // Always clean up SDK state
             _platform = null;
             DefaultLogger = null;
             _logHandler?.Detach();
             _logHandler = null;
             _worker?.Stop();
+            _worker = null;
         }
 
         private void OnQuitting()
         {
-            ShutdownInstance();
+            ShutdownInstance(true);
         }
 
         internal class ConfigKeys